pyproject.toml
[project]
name = "telegram-bot-video-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "ThanhPT",email = "thanhpt@vnpt.vn"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=21.10,<22.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "yt-dlp (>=2025.2.19,<2026.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[tool.ruff]
line-length = 120
target-version = "py312"


[tool.ruff.format]
indent-style = "space"


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]


[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
